- Defined hive_account_operations_account_op_seq_no_id_idx index covering hive_account_operations columns required to efficiently execute AH queries. - Defined hive_accounts_name_idx to allow efficient account by name lookup - hive_operations.op_pos column type changed to integer.

+++ libraries/plugins/sql_serializer/schema.sql
@@ -77,7 +77,7 @@ CREATE TABLE IF NOT EXISTS hive_operations (
   id bigint not null,
   block_num integer NOT NULL,
   trx_in_block smallint NOT NULL,
-  op_pos smallint NOT NULL,
+  op_pos integer NOT NULL,
   op_type_id smallint NOT NULL,
   body text DEFAULT NULL
   --,permlink_ids int[]
